% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining_vvm_life.R
\name{annotate_vvm}
\alias{annotate_vvm}
\title{Add Remaining VVM Life annotations to a temperature time series}
\usage{
annotate_vvm(df)
}
\arguments{
\item{df}{a data frame containing a temperature time series. Must include a \code{date} column containing the datetime, and a \code{temp} column containing the temperature in degrees Celsius.}
}
\value{
the data frame, with extra columns for reaction constants and instantaneous and cumulative VVM life lost
}
\description{
This function adds reaction constants, instantaneous VVM Life Lost, and
cumulative VVM Life Lost columns to a
temperature time series. The calculations are done for the four most common
VVM types:
VVM2, VVM7, VVM14, and VVM30.
}
